Suggestions for what to edit and where for issue41
Issue #41 is slated to be implemented in this document.  I went ahead and implemented it already in the code for BCHN and this is what I suggest you specify here so as to be very clear and avoid subtle consensus errors.

Note that I added the requirement that to PARSE token data correctly, up to 32 MiB compactSize must be supported for COMMITMENT_LENGTH (which is the default limit for all compactSizes in Bitcoin).  Anything above this size fails to PARSE.  However commitments above 40 bytes will fail validation.  This point is important considering that we now have a distinction between:

- correctly parsed token data
- incorrectly parsed token data that just ends up as raw bytes into the locking script as its prefix

Why the 32MiB limit? This is the standard compactSize limit in Bitcoin used for all compactSizes (unless otherwise specified) and I feel it just makes the code do the "natural thing".  *For parsing*, we must accept such crazy token data -- even if it fails consensus later!  This is important because we really need to delineate what is a parse error and what is a consensus error.